LSG-10: create DesktopTable and use in BansTable
